requirements.txt
azure-functions>=1.20.0
azure-ai-vision-imageanalysis>=1.0.0
azure-core>=1.30.0
